+++ src/engine/planning/mod.rs
@@ -1,9 +1,10 @@
 mod filter;
 mod query;
 pub mod query_plan;
+pub mod planner;
 
 pub use self::query_plan::QueryPlan;
-pub use self::query_plan::QueryPlanner;
+pub use self::planner::QueryPlanner;
 pub use self::filter::Filter;
 pub use self::query::Query;
 pub use self::query::NormalFormQuery;
